Moved duplication function from model to api and PR comment changes
